+++ .formatter.exs
@@ -2,6 +2,7 @@
 [
   export: [
     locals_without_parens: [
+      assert_contains: :*,
       assert_recent: :*,
       assert_that: :*,
       refute_that: :*
@@ -10,6 +11,7 @@
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
   line_length: 120,
   locals_without_parens: [
+    assert_contains: :*,
     assert_recent: :*,
     assert_that: :*,
     refute_that: :*